requirements.txt
python-dotenv
pandas
numpy
pymongo
certifi
pymongo[srv]
scikit-learn
pyaml
mlflow
fastapi
uvicorn
scipy
apache-airflow
Jinja2
streamlit
requests 
beautifulsoup4 
python-whois 
apache-airflow-providers-mongo

#-e .